Suppress redundant default parameter value assignment warning
The MapStruct annotation processor makes a difference between a
parameter annotation value with a user defined value and the default value.

For `Mapping#constant` and `Mapping#defaultValue` an empty value
means that MapStruct should use that empty value in the generated code.

With this commit we are using a new extension point added in 2021.2
to ignore IntelliJ warning for the specific annotation parameters

Closes #93